Add faux-websocket/CDN support
We can now use a faux-websocket connection instead of a CONNECT
tunnel. "faux" because the websocket connection is not quite valid, as
framing is not implemented at all, and we treat the connection the
same as a CONNECT tunnel, after the headers are exchanged.

This could be useful in order to access naiveproxy hosted behind a
CDN.